Fix missing/wrong gradle property annotations
Description
===========

Gradle is very picky when it comes to the correct way
of declaring properties for tasks. Our `BaseSpec` uses
`@Inject` getters which are not allowed to also carry `@Internal`
annotations.

I also added the missing annoation for `getArguments` in `ArgumentSpec`

Last but not least I added a validation task from the gradle
`gradle-java-plugin` plugin to validate our mocked plugin/task classes.

Changes
=======

* ![FIX] incorrect property annotations in `BaseSpec`
* ![FIX] missing property annotation for `getArguments` in `ArgumentsSpec`
* ![IMPROVE] gradle property annotiation checks with task `validateMockedPlugins`