Fix ProcessExecutor environment set logic (#26)
Description
===========

The environment handling is broken because of property access
shadowing issue in the closure which passes the saved environment
down to the internal `ExecSpec`. I added a few `this.` to point to
the correct properties.

The tests for this specific case where also broken. There was actually
no test executed. I split the test which tested arguments and environment
setter into two different tests.

I also opted to remove the null check API for the environment setter.
It was either this or allow the arguments list also to be null. I was unable
to write a good test for these cases because of the ambigous overloads when calling
the method with `null`. I thought that only expecting non `null` values is the cleaner
route anyways.

Changes
=======

* ![FIX] `ProcessExecutor` environment set logic