Fix eager evaluation of project.properties in PropertyLookup (#30)
Description
===========

This patch fixes a regression we have since the last patch.
We have a situation where the access to `project.properties` happens
before the base directory is created. This issue happens mainly during
test and can be rooted to nebular test. Or that is my assumption.
I opted to rollback the on line that wraps the access into yet another
`Provider`. It feels unessary but is important.

I also added a commend to make sure we don't reintroduce this, or
can test for potential fix in nebular test.

Changes
=======

![FIX] eager evaluation of `project.properties` in `PropertyLookup`